Moves DoFnAdapters to runners-core
